Reactors weren't removed from standard levels that also contained a boss robot. Clients were disconnected in coop games when trying to rejoin the game and enter the next level after a reactor or boss had been destroyed.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@1324 01821946-851b-0410-b7f4-f1477c70e57a

+++ render/objeffects.cpp
@@ -1679,7 +1679,7 @@ if (!gameData.objs.bIsSlowWeapon [objP->id] && gameStates.app.bHaveExtraGameInfo
 			l = f2fl (phb->vMax.p.z - phb->vMin.p.z);
 			dx = f2fl (phb->vMax.p.x - phb->vMin.p.x);
 			dy = f2fl (phb->vMax.p.y - phb->vMin.p.y);
-			r = (sqrt (dx * dx + dy * dy) / sqrt (2.0f));
+			r = (float) (sqrt (dx * dx + dy * dy) / sqrt (2.0f));
 			if (objP->id == FUSION_ID) {
 				l *= 1.5f;
 				r /= 1.5f;
